requirements.txt
# Django
Django==5.1.5
djangorestframework==3.15.2

# Database
psycopg2-binary==2.9.10

# Security
django-cors-headers==4.6.0
python-decouple==3.8
django-environ==0.11.2
django-axes==6.5.1
django-ratelimit==4.1.0

# Admin enhancements
django-admin-interface==0.28.9
django-colorfield==0.12.0

# API & Authentication
djangorestframework-simplejwt==5.4.0

# Utilities
Pillow==11.1.0
python-dateutil==2.9.0
pytz==2024.2

# Filtering
django-filter==24.3

# Development
django-debug-toolbar==4.4.6
